+++ packages/antwerp-ui/react-components/src/constants/settings.ts
@@ -1,7 +1,7 @@
 import { nlBE } from 'date-fns/locale';
 
 // BASE
-export const ICONS_URL = 'https://cdn.antwerpen.be/core_branding_scss/6.6.0/assets/images/ai.svg';
+export const ICONS_URL = 'https://cdn.antwerpen.be/core_branding_scss/6.6.1/assets/images/ai.svg';
 
 // ATOMS
 export const DEFAULT_AVATAR_ICON = 'single-neutral';