+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Core branding update 6.6.0
+- Core branding update 6.6.1
 
 ## 7.1.1 - 2024-03-07